• Updated Build settings to support 64bit architecture, which triggered a ton of minor Xcode warnings which have been addressed & resolved • Various minor bug fixes

+++ Nooch/DisputeDetail.m
@@ -347,7 +347,7 @@
         [nav_ctrl performSelector:@selector(reset)];
         
         NSMutableArray * arrNav = [nav_ctrl.viewControllers mutableCopy];
-        for (int i = [arrNav count]; i > 1; i--) {
+        for (short i = [arrNav count]; i > 1; i--) {
             [arrNav removeLastObject];
         }
